Expanded index dump to show linked lists.
Laying in code to remove a row from all indexes but i need a fast way to
see if an index exists or not based on index number, not field number.

About to go through and change all the places where it indexes into skipLists
by field to do it by index.

Going to have a bit in the field structure for a table that says if it
is indexed.