Make skiplist search selection table-driven. Fix a couple of symptomless typos (eg: missing break before end of switch). Make < and >= use skiplists. Update tests to match.
Note: there should be a way to make it NOT use skiplists, when you know that
it's going to hit most of the rows and so a hashtable is going to be faster.

Either "-index false" or a new option indicating that you expect to hit
most or all of the table.