Laying out structured for multi-reader one-writer shared pool.
We are assuming that skiplists can be traversed live.

Every write, the writer increments a "cycle" in the map.
The reader copies that "cycle" when it starts doing something with the
ctable, and clears its copy when it is done. It copies EVERYTHING it's
gong to return to Tcl out of shared memory before it clears it.

When the writer frees memory it puts it on an "unfreed" list of pending
frees. Periodically it goes through and actually frees for reuse the
blocks belonging to readers who have a lower cycle or a zero cycle.