Written ckalloc-alike wrappor for shmalloc.
Written code to generate two possible sets of allocators, depending on whether shared table support is enabled or not.

Hash will be in memory for master only. skiplists will be in everyone's
space. TODO: in generating tables, if there are no skiplists it's an error
to create a master or reader table! Should be possible to disable with no
skiplists.

Written code to generate an in-core base row for hash lookups in ctables,
the rest of the table will be in shared memory. TODO: in generating tables,
the "key" column MUST be the default _key for shared tables. This is determined
early enough that we can switch to the non-shared allocator and disable the
commands to create shared tables in this case.