New "performance_callback" method for tables
* $tableName performance_callback 0.001 my_callback

When invoked in this manner, if any speedtable search takes more than 0.001 user CPU seconds,
my_callback is performed with three arguments: the speedtables search command, the number of
rows selected, and the user CPU seconds.

* $tableName performance_callback

When invoked in this manner, a list consisting of the current cpu callback threshold and the
callback routine is returned.

* $tableName 0 ""

When invoked in this manner, the callback function is disabled.  Any value can be passed for
CPU time, it's the empty callback function what switches it off.